renamed internals to internal
